Move MakieCore to an extension
