+++ GeoInterfaceMakie/Project.toml
@@ -6,7 +6,12 @@ version = "0.1.9"
 [deps]
 GeoInterface = "cf35fbd7-0cd7-5166-be24-54bfbe79505f"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
+
+[weakdeps]
 MakieCore = "20f20a25-4f0e-4fdf-b5d1-57303727442b"
+
+[extensions]
+GeoInterfaceMakieMakieCoreExt = "MakieCore"
 
 [compat]
 GeoInterface = "1"
